Fix error in ipadrop & optimize nuke.
